Move run notifier classes to Common directory
Re #23027
