+++ qt/scientific_interfaces/ISISReflectometry/GUI/Common/CMakeLists.txt
@@ -1,9 +1,12 @@
 set ( COMMON_SRC_FILES
+  CatalogRunNotifier.cpp
   Plotter.cpp
 )
 
 # Include files aren't required, but this makes them appear in Visual Studio
 set ( COMMON_INC_FILES
+  CatalogRunNotifier.h
+  IRunNotifier.h
   IPlotter.h
   Plotter.h
 )
