+++ qt/scientific_interfaces/ISISReflectometry/GUI/Runs/RunsPresenter.cpp
@@ -9,8 +9,8 @@
 #include "CatalogSearcher.h"
 #include "Common/IMessageHandler.h"
 #include "GUI/Batch/IBatchPresenter.h"
+#include "GUI/Common/IRunNotifier.h"
 #include "GUI/RunsTable/RunsTablePresenter.h"
-#include "IRunNotifier.h"
 #include "IRunsView.h"
 #include "MantidAPI/AlgorithmManager.h"
 #include "MantidAPI/CatalogManager.h"
